refactor: extract shared CommandExecutorMixin and simplify functions
Genuine improvements to reduce code duplication and complexity:

1. Extract CommandExecutorMixin to modes/__init__.py
   - Shared _execute_commands() was 100% duplicated between browser.py and terminal.py
   - Extracted narration timing helpers (_handle_narration_before/after)
   - BrowserRecorder and TerminalRecorder now inherit from mixin

2. Simplify _execute_vim_sequence in terminal.py
   - Replace if/elif chain with dict dispatch for special keys
   - Reduces from 11 to 9 lines

3. Reduce verbose docstring in stage.py::calculate_visible_range
   - 6-line docstring → 1-line docstring (params are self-documenting)

Results:
- 12-line functions: 13 → 11 (-2)
- 11-line functions: 7 → 6 (-1)
- Total warnings: 49 → 47 (-2)
- Lines of code: -28 net reduction
